+++ src/activity/Activity.js
@@ -113,6 +113,7 @@ export function Activity(Behaviour, activityDef, context) {
     isAdHoc: activityDef.isAdHoc,
     isParallelJoin,
     isParallelGateway: activityDef.isParallelGateway,
+    isConvergingGateway: activityDef.isConvergingGateway,
     isStartEvent: !!activityDef.isStartEvent,
     isThrowing: activityDef.isThrowing,
     linkNames: activityDef.linkNames,
@@ -253,6 +254,11 @@ Object.defineProperties(Activity.prototype, {
   isParallelGateway: {
     get() {
       return this[K_FLAGS].isParallelGateway;
+    },
+  },
+  isConvergingGateway: {
+    get() {
+      return this[K_FLAGS].isConvergingGateway;
     },
   },
   isStartEvent: {
@@ -648,7 +654,7 @@ Activity.prototype._discardRun = function discardRun() {
 
 /** @internal */
 Activity.prototype._onShakeMessage = function _onShakeMessage(sourceMessage) {
-  if (this[K_FLAGS].isParallelGateway) {
+  if (this[K_FLAGS].isConvergingGateway) {
     const message = cloneMessage(sourceMessage, { join: this.id });
     message.content.sequence.push({ id: this.id, type: this.type });
     return this.broker.publish('event', 'activity.shake.converge', message.content, {